renamed files: "boardlayerprovider.cpp/h" --> "boardlayerstack.cpp/h"
